refactor(meetup): improve meetup-related with better syntax and structure
