+++ src/main/kotlin/com/trailoapp/trailo_backend/domain/enum/geo/MeetupStatus.kt
@@ -1,4 +1,4 @@
-package com.trailoapp.trailo_backend.domain.enum
+package com.trailoapp.trailo_backend.domain.enum.geo
 
 enum class MeetupStatus {
     WAITING, CURRENT, CANCELLED, FINISHED
