+++ src/main/kotlin/com/trailoapp/trailo_backend/domain/enum/geo/TerrainType.kt
@@ -1,4 +1,4 @@
-package com.trailoapp.trailo_backend.domain.enum
+package com.trailoapp.trailo_backend.domain.enum.geo
 
 enum class TerrainType {
     UNSPECIFIED, ROCKY, MUDDY, SANDY, FOREST, MOUNTAIN, DESERT, RIVER, SNOW, GRAVEL, CLAY, VOLCANIC
