+++ src/main/kotlin/com/trailoapp/trailo_backend/domain/enum/social/FriendshipStatus.kt
@@ -1,4 +1,4 @@
-package com.trailoapp.trailo_backend.domain.enum
+package com.trailoapp.trailo_backend.domain.enum.social
 
 enum class FriendshipStatus {
     PENDING, ACCEPTED, REJECTED
