+++ src/main/kotlin/com/trailoapp/trailo_backend/dto/meetup/request/UpdateMeetupRequest.kt
@@ -1,8 +1,8 @@
 package com.trailoapp.trailo_backend.dto.meetup.request
 
-import com.trailoapp.trailo_backend.domain.enum.MeetupStatus
-import com.trailoapp.trailo_backend.domain.enum.TerrainType
-import com.trailoapp.trailo_backend.domain.enum.TrailDifficulty
+import com.trailoapp.trailo_backend.domain.enum.geo.MeetupStatus
+import com.trailoapp.trailo_backend.domain.enum.geo.TerrainType
+import com.trailoapp.trailo_backend.domain.enum.geo.TrailDifficulty
 import java.math.BigDecimal
 import java.time.OffsetDateTime
