+++ src/main/kotlin/com/trailoapp/trailo_backend/service/GroupService.kt
@@ -1,8 +1,8 @@
 package com.trailoapp.trailo_backend.service
 
 import com.trailoapp.trailo_backend.domain.core.UserEntity
-import com.trailoapp.trailo_backend.domain.enum.GroupRoles
-import com.trailoapp.trailo_backend.domain.enum.MembershipStatus
+import com.trailoapp.trailo_backend.domain.enum.social.GroupRoles
+import com.trailoapp.trailo_backend.domain.enum.social.MembershipStatus
 import com.trailoapp.trailo_backend.domain.core.GroupEntity
 import com.trailoapp.trailo_backend.domain.social.UserGroupEntity
 import com.trailoapp.trailo_backend.dto.group.request.CreateGroupRequest
